Revert "Adds TypeScript (#10)"
This reverts commit e0aa04649d3b55f0f9e60e81347c863bd74562d3.
