+++ client/package.json
@@ -3,10 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "^26.0.24",
-    "@types/node": "^16.3.3",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
     "ethers": "^5.4.1",
@@ -34,8 +30,6 @@
     ]
   },
   "devDependencies": {
-    "prettier": "2.3.2",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.3.5"
+    "prettier": "2.3.2"
   }
 }
