+++ client/src/App.js
@@ -1,10 +1,9 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
+import React from "react";
+import Web3Provider from "./Web3Provider";
 import Web3Connect from "./Web3Connect";
 import Web3App from "./Web3App";
-import Web3Provider from "./Web3Provider";
 
-const App = () : JSX.Element => {
+const App = () => {
   return (
     <div className="App">
       <Web3Provider>
