+++ client/src/Web3Provider.js
@@ -1,7 +1,7 @@
 import { Web3ReactProvider } from "@web3-react/core";
 import { Web3Provider as EthersJsWeb3Provider } from "@ethersproject/providers";
 
-const getLibrary = (provider) : EthersJsWeb3Provider => {
+const getLibrary = (provider) => {
   const lib = new EthersJsWeb3Provider(provider);
   lib.pollingInterval = 1000;
   return lib;
