+++ client/src/components/RecordKit/RecordKitForm.js
@@ -1,12 +1,68 @@
 import React, { useRef, useState, useEffect } from "react";
 import { useCreateRecord } from "../../hooks/useRecords";
-import useInputForm  from "../../hooks/useInputForm";
 
-interface IRecordKitFormProps {
-	 labelId : string
-}
+const useRecordKitForm = ({ initialValues, onSubmit }) => {
+  const [values, setValues] = useState(initialValues || {});
+  const [errors, setErrors] = useState({});
+  const [touched, setTouched] = useState({});
+  const [onSubmitting, setOnSubmitting] = useState(false);
+  const [onBlur, setOnBlur] = useState(false);
 
-const RecordKitForm = ({ labelId } : IRecordKitFormProps) => {
+  const formRendered = useRef(true);
+
+  useEffect(() => {
+    if (formRendered.current) {
+      setValues(initialValues);
+      setErrors({});
+      setTouched({});
+      setOnSubmitting(false);
+      setOnBlur(false);
+    }
+    formRendered.current = false;
+  }, [initialValues]);
+
+  const handleBlur = (event) => {
+    const { target } = event;
+    const { name } = target;
+    setTouched({ ...touched, [name]: true });
+    setErrors({ ...errors });
+  };
+
+  // Set any errors if applicable
+  const handleChange = (event) => {
+    const { target } = event;
+    const { name, value } = target;
+    event.persist();
+    setValues({ ...values, [name]: value });
+  };
+
+  // Combines the file values from upload field to the values
+  const handleFileChange = (event) => {
+    const { target } = event;
+    const { name, files } = target;
+    event.persist();
+    setValues({ ...values, [name]: files[0] });
+  };
+
+  // Set any errors if applicable and submit with onSubmit
+  const handleSubmit = (event) => {
+    if (event) event.preventDefault();
+    setErrors({ ...errors });
+    onSubmit({ values, errors });
+  };
+
+  return {
+    values,
+    errors,
+    touched,
+    handleChange,
+    handleFileChange,
+    handleBlur,
+    handleSubmit,
+  };
+};
+
+const RecordKitForm = ({ labelId }) => {
   const [record, setRecord] = useState(null);
   const [{ data, isLoading, error }, createRecord] = useCreateRecord();
 
@@ -29,7 +85,7 @@ const RecordKitForm = ({ labelId } : IRecordKitFormProps) => {
     handleFileChange,
     handleBlur,
     handleSubmit,
-  } = useInputForm({
+  } = useRecordKitForm({
     initialValues,
     onSubmit: onUpload,
   });
