--- client/src/hooks/useMint.ts
@@ -1,8 +1,0 @@
-import React, { useState, useEffect } from "react";
-
-export const useMint = () => {
-  const ethersjsInstance = useEthersJs();
-	// TODO 
-}
-
-
