+++ client/src/hooks/useRecords.js
@@ -2,7 +2,7 @@ import React, { useCallback, useState, useEffect } from "react";
 import { useRecordRepository } from "../hooks/useRepository";
 import { REC_HASH_RATE } from "../model/Record";
 
-export const useRecords = (labelId) : Array<Record> => {
+export const useRecords = (labelId) => {
   const [records, setRecords] = useState(null);
   const repo = useRecordRepository();
 
@@ -21,21 +21,7 @@ export const useRecords = (labelId) : Array<Record> => {
   return records;
 };
 
-type CreateRecordResult = {
-	 data : any;
-	 error : any;
-	 isLoading : boolean;
-};
-
-interface IRecordDraft {
-	 title : string;
-	 artist : string;
-	 audio : File;
-}
-
-type CreateRecordRequest = (labelId : string, values : IDraftRecord) => void;
-
-export const useCreateRecord : Array<CreateRecordResult, CreateRecordRequest> = () => {
+export const useCreateRecord = () => {
   const [result, setResult] = useState({
     data: null,
     error: null,
@@ -44,6 +30,7 @@ export const useCreateRecord : Array<CreateRecordResult, CreateRecordRequest> = 
 
   const repo = useRecordRepository();
 
+  // You POST method here
   const request = (labelId, record) => {
     setResult((prev) => ({ ...prev, isLoading: true }));
